+++ rs/crypto/internal/crypto_service_provider/src/vault/local_csp_vault/tschnorr/tests.rs
@@ -14,7 +14,7 @@ use ic_crypto_internal_threshold_sig_canister_threshold_sig::{
 };
 use ic_crypto_test_utils_reproducible_rng::reproducible_rng;
 use ic_types::{
-    crypto::{canister_threshold_sig::ExtendedDerivationPath, AlgorithmId},
+    crypto::{AlgorithmId, ExtendedDerivationPath},
     Randomness,
 };
 use proptest::{
