+++ rs/crypto/utils/canister_threshold_sig/src/lib.rs
@@ -1,8 +1,7 @@
 use ic_crypto_internal_threshold_sig_canister_threshold_sig::DeriveThresholdPublicKeyError;
 use ic_types::crypto::canister_threshold_sig::error::CanisterThresholdGetPublicKeyError;
-use ic_types::crypto::canister_threshold_sig::{
-    ExtendedDerivationPath, MasterPublicKey, PublicKey,
-};
+use ic_types::crypto::canister_threshold_sig::{MasterPublicKey, PublicKey};
+use ic_types::crypto::ExtendedDerivationPath;
 
 /// Derives the threshold public key from the specified `master_public_key` for
 /// the given `extended_derivation_path`.
