+++ rs/types/types/src/consensus/idkg/schnorr.rs
@@ -1,7 +1,8 @@
 //! Threshold Schnorr transcripts and references related definitions.
 use crate::crypto::canister_threshold_sig::{
-    ExtendedDerivationPath, SchnorrPreSignatureTranscript, ThresholdSchnorrSigInputs,
+    SchnorrPreSignatureTranscript, ThresholdSchnorrSigInputs,
 };
+use crate::crypto::ExtendedDerivationPath;
 use crate::{Height, Randomness};
 #[cfg(test)]
 use ic_exhaustive_derive::ExhaustiveSet;
